Merge in SLI/itest from ~NXF97358/ele_nikhil:itest_ele to itest_ele

* commit '4a1762e1eb00ae6f77b877c935aec10381fa439e':
  LF-11697: Update time variables used in calculation

+++ inc/itest.h
@@ -129,12 +129,12 @@ while(0)
 typedef struct {
     struct timespec ts1; // for total iterations
     struct timespec ts2; // for total iterations
-    uint64_t time_us;
-    uint64_t min_time_us;
-    uint64_t max_time_us;
+    double time_us;
+    double min_time_us;
+    double max_time_us;
     uint32_t nb_iter;
     uint32_t op_sec;
-    uint32_t t_per_op;
+    double t_per_op;
 } timer_perf_t;
 
 typedef struct{
@@ -171,7 +171,7 @@ void init_timer(timer_perf_t *timer);
 void start_timer(timer_perf_t *timer);
 void stop_timer(timer_perf_t *timer);
 void finalize_timer(timer_perf_t *timer, uint32_t nb_iter);
-uint64_t timespec_elapse_usec(struct timespec *ts1, struct timespec *ts2);
+double timespec_elapse_usec(struct timespec *ts1, struct timespec *ts2);
 void print_perf(timer_perf_t *timer);
 
 #endif
